# frontend/src/lib/openapi.json
{
  "components": {
    "schemas": {
      "ActionDTO": {
        "additionalProperties": false,
        "properties": {
          "argument": {
            "default": "",
            "maxLength": 4096,
            "title": "Argument",
            "type": "string"
          },
          "type": {
            "enum": [
              "fileinto",
              "fileinto_copy",
              "redirect",
              "addflag",
              "reject",
              "keep",
              "discard",
              "stop"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type"
        ],
        "title": "ActionDTO",
        "type": "object"
      },
...
      "Body_import_script_api_scripts_import_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "file"
        ],
        "title": "Body_import_script_api_scripts_import_post",
        "type": "object"
      },
      "ConditionDTO": {
        "additionalProperties": false,
        "properties": {
          "address_part": {
            "default": "",
            "enum": [
              "",
              "all",
              "localpart",
              "domain"
            ],
            "title": "Address Part",
            "type": "string"
          },
          "address_test": {
            "default": false,
            "title": "Address Test",
            "type": "boolean"
          },
          "comparator": {
            "default": "",
            "maxLength": 80,
            "title": "Comparator",
            "type": "string"
          },
          "header": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Header",
            "type": "string"
          },
          "match_type": {
            "enum": [
              "contains",
              "is",
              "matches",
              "regex"
            ],
            "title": "Match Type",
            "type": "string"
          },
          "negate": {
            "default": false,
            "title": "Negate",
            "type": "boolean"
          },
          "value": {
            "default": "",
            "maxLength": 4096,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "header",
          "match_type"
        ],
        "title": "ConditionDTO",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OkResponse": {
        "description": "Generic mutating-endpoint response. `name` / `sieve` are populated\nby individual routes; `response_model_exclude_none=True` keeps the\npayload minimal.",
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "sieve": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sieve"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "title": "OkResponse",
        "type": "object"
      },
      "PreviewRequest": {
        "additionalProperties": false,
        "description": "One Rule to render as Sieve (areyousievious-8fg.18 vocabularies apply).\n\nDeliberately one Rule and not a whole script: the editor previews the rule\nthe user is looking at, and asking for a script would make the endpoint\nneed `requires`, which is a property of the script rather than of any Rule.",
        "properties": {
          "rule": {
            "$ref": "#/components/schemas/RuleDTO"
          }
        },
        "required": [
          "rule"
        ],
        "title": "PreviewRequest",
        "type": "object"
      },
      "PreviewResponse": {
        "description": "POST /api/scripts/preview \u2014 the exact bytes a save would write for this\nRule, minus the script-level `require` line.",
        "properties": {
          "sieve": {
            "title": "Sieve",
            "type": "string"
          }
        },
        "required": [
          "sieve"
        ],
        "title": "PreviewResponse",
        "type": "object"
      },
      "RawBlockDTO": {
        "additionalProperties": false,
        "description": "A Raw Block entry on the wire \u2014 Sieve the parser didn't recognise.",
        "properties": {
          "comment": {
            "default": "",
            "maxLength": 4096,
            "title": "Comment",
            "type": "string"
          },
          "kind": {
            "const": "raw",
            "default": "raw",
            "title": "Kind",
            "type": "string"
          },
          "source": {
            "default": "",
            "maxLength": 65536,
            "title": "Source",
            "type": "string"
          },
          "text": {
            "default": "",
            "maxLength": 65536,
            "title": "Text",
            "type": "string"
          }
        },
        "title": "RawBlockDTO",
        "type": "object"
      },
      "RuleDTO": {
        "additionalProperties": false,
        "description": "A Rule entry on the wire.\n\nCarries no `id` \u2014 identity is view state (docs/adr/0001-identity-is-view-state.md).\nClients mint their own render keys and strip them before saving.",
        "properties": {
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionDTO"
            },
            "maxItems": 64,
            "title": "Actions",
            "type": "array"
          },
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/ConditionDTO"
            },
            "maxItems": 64,
            "title": "Conditions",
            "type": "array"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "kind": {
            "const": "rule",
            "default": "rule",
            "title": "Kind",
            "type": "string"
          },
          "match": {
            "default": "anyof",
            "enum": [
              "",
              "anyof",
              "allof"
            ],
            "title": "Match",
            "type": "string"
          },
          "name": {
            "default": "",
            "maxLength": 200,
            "title": "Name",
            "type": "string"
          },
          "source": {
            "default": "",
            "maxLength": 65536,
            "title": "Source",
            "type": "string"
          }
        },
        "title": "RuleDTO",
        "type": "object"
      },
      "SaveRawRequest": {
        "additionalProperties": false,
        "description": "`content` is required and non-empty: an empty raw save destroys the\nuser's script, and the one observed way to send one was a bug, not an\nintent (areyousievious-8fg.20). Clearing a script is expressed by\ndeleting it, not by overwriting it with nothing.",
        "properties": {
          "content": {
            "maxLength": 262144,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "title": "SaveRawRequest",
        "type": "object"
      },
      "SaveScriptRequest": {
        "additionalProperties": false,
        "description": "Typed save body. Every entry is parsed against its DTO and an unknown\nfield anywhere in the tree is a hard 422.\n\n`entries` is one ordered sequence \u2014 position is the evaluation order. It\nreplaced parallel rules/raw_blocks/order arrays, where an order array that\nomitted one of its own rules caused that rule to be silently dropped.",
        "properties": {
          "entries": {
            "items": {
              "discriminator": {
                "mapping": {
                  "raw": "#/components/schemas/RawBlockDTO",
                  "rule": "#/components/schemas/RuleDTO"
                },
                "propertyName": "kind"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/RuleDTO"
                },
                {
                  "$ref": "#/components/schemas/RawBlockDTO"
                }
              ]
            },
            "maxItems": 1000,
            "title": "Entries",
            "type": "array"
          },
          "preamble": {
            "default": "",
            "maxLength": 65536,
            "title": "Preamble",
            "type": "string"
          },
          "requires": {
            "items": {
              "type": "string"
            },
            "maxItems": 64,
            "title": "Requires",
            "type": "array"
          },
          "requires_source": {
            "default": "",
            "maxLength": 8192,
            "title": "Requires Source",
            "type": "string"
          },
          "tail": {
            "default": "",
            "maxLength": 65536,
            "title": "Tail",
            "type": "string"
          }
        },
        "title": "SaveScriptRequest",
        "type": "object"
      },
...
      "ScriptRawResponse": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "content"
        ],
        "title": "ScriptRawResponse",
        "type": "object"
      },
      "ScriptResponse": {
        "description": "GET /api/scripts/{name} \u2014 mirrors what `script_to_json` emits.",
        "properties": {
          "entries": {
            "items": {
              "discriminator": {
                "mapping": {
                  "raw": "#/components/schemas/RawBlockDTO",
                  "rule": "#/components/schemas/RuleDTO"
                },
                "propertyName": "kind"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/RuleDTO"
                },
                {
                  "$ref": "#/components/schemas/RawBlockDTO"
                }
              ]
            },
            "title": "Entries",
            "type": "array"
          },
          "preamble": {
            "default": "",
            "maxLength": 65536,
            "title": "Preamble",
            "type": "string"
          },
          "requires": {
            "items": {
              "type": "string"
            },
            "title": "Requires",
            "type": "array"
          },
          "requires_source": {
            "default": "",
            "maxLength": 8192,
            "title": "Requires Source",
            "type": "string"
          },
          "tail": {
            "default": "",
            "maxLength": 65536,
            "title": "Tail",
            "type": "string"
          }
        },
        "title": "ScriptResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AreYouSievious",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/scripts/import": {
      "post": {
        "description": "Import a .sieve file as a new script.\n\nponytail: sync handler so FastAPI runs it in a threadpool \u2014 the slow\nManageSieve PUT no longer blocks the event loop (Perf C1 / Fwk C-1).",
        "operationId": "import_script_api_scripts_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_script_api_scripts_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Script"
      }
    },
    "/api/scripts/preview": {
      "post": {
        "description": "Render one Rule as the house-style Sieve an EDIT to it would write.\n\nSince areyousievious-8fg.14 a save re-emits an unedited Rule's original\nbytes and regenerates only what changed, so this is not the text every save\nwrites \u2014 it is the text this Rule takes once it is edited. That is what\nmakes it worth showing: the reformatting is the part the user has not\ncommitted to, and they see it before they do. A Rule they leave alone keeps\nthe bytes it came with.\n\nNO MAIL-SERVER DIAL. It depends on `get_session` and not on\n`get_script_store`, so it authenticates without opening a ManageSieve\nconnection \u2014 this runs on every keystroke behind a debounce, and a\nconnection per keystroke would be a self-inflicted denial of service\nagainst the user's own mail server.\n\nIt replaces `previewRule` in the SPA, which was a second implementation of\n`SieveGenerator` that had already diverged five ways. That sharing is still\nreal: preview and the regenerating save path go through one\n`generate_entry`, so they cannot drift again. Declared BEFORE the `/{name}`\nroutes so `preview` is read as a literal path segment.",
        "operationId": "preview_rule_api_scripts_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Rule"
      }
    },
    "/api/scripts/{name}": {
      "delete": {
        "operationId": "delete_script_api_scripts__name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Script"
      },
      "get": {
        "description": "Get script parsed as JSON rules.",
        "operationId": "get_script_api_scripts__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScriptResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Script"
      },
      "put": {
        "description": "Save script from JSON rules (generates Sieve).\n\nPRE-FLIGHT (areyousievious-8fg.13): the Sieve we generated is checked\nagainst an independent grammar before it is sent. A Rule whose last\nCondition was deleted generates `if anyof ( ) {`, which a real server\nrefuses \u2014 the error contract already made that rejection honest, and this\nmeans we never send it at all.\n\nONLY the Rules are checked, never the RawBlocks. sievelib\ndoes not know `include`, `addheader` or `spamtest` though real servers do,\nso validating the whole script would refuse working scripts over a\nconstruct we never touched.\n\nTHE SAME PRE-FLIGHT ALSO GUARDS THE VERBATIM BOUNDARY BYTES\n(areyousievious-8fg.14). `preamble`, `requires_source` and `tail` arrive on\nthe wire and are written to the mail server unaltered, and unlike an\nentry's `source` they have no entry to be compared against \u2014 so\n`preflight_error` additionally refuses a head that parses to anything but\nthe requires we declare, and a tail that carries a statement at all. This\nis the only call site: one request, one gate.",
        "operationId": "save_script_api_scripts__name__put",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveScriptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Script"
      }
    },
    "/api/scripts/{name}/activate": {
      "post": {
        "operationId": "activate_script_api_scripts__name__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activate Script"
      }
    },
    "/api/scripts/{name}/export": {
      "get": {
        "description": "Download script as a .sieve file.",
        "operationId": "export_script_api_scripts__name__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Script"
      }
    },
    "/api/scripts/{name}/raw": {
      "get": {
        "description": "Get raw Sieve text.",
        "operationId": "get_script_raw_api_scripts__name__raw_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScriptRawResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Script Raw"
      },
      "put": {
        "description": "Save raw Sieve text directly.",
        "operationId": "save_script_raw_api_scripts__name__raw_put",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveRawRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Script Raw"
      }
    },
...
  }
}